refactor: VO 패키지 분리
